kvserver: move LoadAndReconcileReplicas to kvstorage
Mechanical using GoLand.

Next steps are starting to add a datadriven testing framework
so that functionality can be exercised when it moves or is added here,
i.e. https://github.com/cockroachdb/cockroach/issues/93247.

Epic: none
Release note: None
