+++ pkg/kv/kvserver/BUILD.bazel
@@ -134,6 +134,7 @@ go_library(
         "//pkg/kv/kvserver/kvadmission",
         "//pkg/kv/kvserver/kvserverbase",
         "//pkg/kv/kvserver/kvserverpb",
+        "//pkg/kv/kvserver/kvstorage",
         "//pkg/kv/kvserver/liveness",
         "//pkg/kv/kvserver/liveness/livenesspb",
         "//pkg/kv/kvserver/logstore",
@@ -354,6 +355,7 @@ go_test(
         "//pkg/kv/kvserver/intentresolver",
         "//pkg/kv/kvserver/kvserverbase",
         "//pkg/kv/kvserver/kvserverpb",
+        "//pkg/kv/kvserver/kvstorage",
         "//pkg/kv/kvserver/liveness",
         "//pkg/kv/kvserver/liveness/livenesspb",
         "//pkg/kv/kvserver/logstore",
